Add support for XLSX
xlrd added support for XLSX with version 0.8.0. This commit declares
the required version of xlrd to be >=0.8.0.

Since the XLS and XLSX support is provided by the same library, the
code for both formats is the same and it therefore doesn't make
sense to have seperate classes for the two fileformats. And so, the
Xls class has been renamed Excel.